# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for managing assignments, questions, classrooms and more.",
        "title": "ParableMind API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/msqs": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MSQs"
                ],
                "summary": "Get All MSQs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by bank ID",
                        "name": "bankId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/questions.MSQ"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MSQs"
                ],
                "summary": "Create MSQ",
                "parameters": [
                    {
                        "description": "MSQ JSON",
                        "name": "msq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/questions.MSQ"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/questions.MSQ"
                        }
                    }
                }
            }
        },
        "/msqs/bulk": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MSQs"
                ],
                "summary": "Bulk Create MSQs",
                "parameters": [
                    {
                        "description": "List of MSQs",
                        "name": "msqs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/questions.MSQ"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/questions.MSQ"
                            }
                        }
                    }
                }
            }
        },
        "/msqs/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MSQs"
                ],
                "summary": "Get MSQ by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MSQ ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/questions.MSQ"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MSQs"
                ],
                "summary": "Update MSQ",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MSQ ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated MSQ",
                        "name": "msq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/questions.MSQ"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/questions.MSQ"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "MSQs"
                ],
                "summary": "Delete MSQ",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MSQ ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MSQs"
                ],
                "summary": "Patch MSQ",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MSQ ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "updates",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "questions.MSQ": {
            "type": "object",
            "properties": {
                "answerIndices": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "bankId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "points": {
                    "type": "integer"
                },
                "question": {
                    "type": "string"
                },
                "variableIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    }
}